backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
keras==3.5.0
pillow==10.4.0
requests==2.32.3
tensorflow==2.17.0
streamlit
groq
pillow
ultralytics
streamlit-chat
